# docs/openapi.yaml
openapi: 3.0.0
info:
  title: File Vault (Safehouse) API
  version: 1.0.0
  description: |
    A secure, cloud-native file vault with AES-256-GCM encryption and AWS S3 storage.
    Feature-flag integrated (via SDK) - flag driven behaviors.
servers:
  - url: http://localhost:4000/api
    description: Local dev
  - url: https://api.file-vault-demo.com/api
    description: Deployed
...
paths:
...
  /vaults:
    get:
      tags: [Vaults]
      summary: List vaults for the authenticated user
      security: [ { bearerAuth: [] } ]
      responses:
        '200':
          description: Array of vaults
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Vault' }
        '401':
          description: Missing/invalid JWT
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    post:
      tags: [Vaults]
      summary: Create a new vault
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVaultRequest'
      responses:
        '201':
          description: Vault created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vault'
        '401':
          description: Missing/invalid JWT
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    CreateVaultRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, example: Demo Vault }
    Vault:
      type: object
      properties:
        id: { type: string, example: 663a3f0d2a... }
        name: { type: string, example: Demo Vault }
        owner: { type: string, example: 663a3f0d2a... }
        createdAt: { type: string, format: date-time }
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string, example: Unauthorized }